file: 주문 상세 조회 컴포넌트 분리
